feat(search): add actions on "search" views
Currently these are the tag and feed lists of articles. We can add/remove tags and do the same actions on the list as on individual articles.

To prevent high RAM consumption with association or dissociation of tags with lots of articles, we add a paginator to do the operation in bulk.