feat(articles): improve pagination
We remove infinite scrolling: it cannot work correctly with read on scroll because the list of articles in the paginator is changing as we load pages. And on other pages, it's easier to apprend the number of pages and where we are with a standard paginator.

We use a clean elided page range.

To prevent having multiple pages with read on scroll, we increase the number of articles per page.

For pages with read on scroll and too many articles for a page, we display a message asking for a page refresh to view the rest.