fix(feeds): rely on the full date and not the day to filter previous runs
It worked for feeds updated once a week since, for example, Mondays will have a different day number each week. But for feeds to update once a month, the previous run was always on the same day number. Using the full date will solve this for all cases.

No idea why it even was implemented like this in the first place.

Close #421