Revert "chore: install django-minify-html"
This reverts commit 367839468adb0f97b70ea53425777ac3b873d55f.

HTMX actions are broken when HTML is minified. It comes from the fact that type="submit" is stripped because it’s the default value for buttons. But HTMX needs it to find the submitter and read formaction attribute to find the proper URL. So the behavior breaks our workflow. I couldn’t find an option to turn that off. Tied to https://github.com/wilsonzlin/minify-html/issues/190

Let’s keep the commit to have a trace we used it and find the project name and its configuration easily again in the future.