fix(browser): try to save URL of the actual page in reader mode
In FF, when reader mode is active, the page content is inaccessible and the URL is hidden because in this mode the page is rendered on a special page extensions cannot access. It’s still possible to fetch the page URL. Getting the content will always trigger an error. So:

- Prevent fatal error if the extension fails to get the content.
- If no content or title is supplied, only post the URL to prevent a 422 error from the backend.
- Parse the internal about:reader URL to extract the actual page URL.

Close #423