fix(api): correct access to documentation
The object we used to identify unset fields in our Pydantic schemas for partial updates wasn’t serializable. This prevented the documentation from generating its examples. Because of this, the documentation page was unusable.

We now use a proper object. We can check whether a field is unset by checking the type of the value. This object takes a function that will provide the example value to the user.