fix(reading): hide the HTML select when updating the page with HTMX boost
The issue was visible on the article details page when updating tags. You could view the tag selector and the raw HTML select. Everything was fine on the initial page load.

Instead of using class= in the HTML which caused two class attributes to be added to the element, we now merge the classes in the code when computing the attributes.